♻️ Move RSS item construction to custom inits
